move tagger to separate project add infra and domain projects
