+++ ImageHosting.Storage.Domain/Entities/ForbiddenCategory.cs
@@ -1,0 +1,6 @@
+namespace ImageHosting.Storage.Domain.Entities;
+
+public class ForbiddenCategory
+{
+    public string Name { get; set; } = null!;
+}
