+++ ImageHosting.Storage.Infrastructure/Converters/UserIdConverter.cs
@@ -1,0 +1,6 @@
+using ImageHosting.Storage.Domain.ValueTypes;
+using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
+
+namespace ImageHosting.Storage.Infrastructure.Converters;
+
+public class UserIdConverter() : ValueConverter<UserId, Guid>(userId => userId.Id, guid => new UserId(guid));
