+++ ImageHosting.Storage.Tagger/AssignTagsConsumer.cs
@@ -1,0 +1,20 @@
+using ImageHosting.Storage.Domain.Messages;
+using ImageHosting.Storage.Domain.Services;
+using MassTransit;
+
+namespace ImageHosting.Storage.Tagger;
+
+public class AssignTagsConsumer(IAssignTagsService assignTagsService) : IConsumer<Batch<CategorizedNewImage>>
+{
+    public Task Consume(ConsumeContext<Batch<CategorizedNewImage>> context)
+    {
+        var messages = context.Message
+            .GroupBy(consumeContext => consumeContext.Message.ImageId,
+                consumeContext => consumeContext.Message.Categories)
+            .ToDictionary(grouping => grouping.Key,
+                grouping => grouping.SelectMany(categories => categories)
+                    .ToDictionary(category => category.Key, category => category.Value));
+
+        return assignTagsService.AssignTagsAsync(messages, context.CancellationToken);
+    }
+}
