+++ ImageHosting.Storage.Tagger/appsettings.json
@@ -1,0 +1,22 @@
+{
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft.AspNetCore": "Warning",
+      "Microsoft.EntityFrameworkCore": "Warning"
+    }
+  },
+  "AssignTags": {
+    "Threshold": 0.8
+  },
+  "AssignTagsConsumer": {
+    "BootstrapServers": [
+      "localhost:9094"
+    ],
+    "TopicName": "categories.v1",
+    "GroupId": "image-tagger"
+  },
+  "ConnectionStrings": {
+    "ImageHosting": "Host=localhost;Port=5432;Database=image_hosting;Username=image_hosting;Password=P@ssw0rd;Include Error Detail=true"
+  }
+}
