--- ImageHosting.Storage/AssignTagsConsumerDefinition.cs
@@ -1,14 +1,0 @@
-using System;
-using MassTransit;
-
-namespace ImageHosting.Storage;
-
-public class AssignTagsConsumerDefinition : ConsumerDefinition<AssignTagsConsumer>
-{
-    protected override void ConfigureConsumer(IReceiveEndpointConfigurator endpointConfigurator,
-        IConsumerConfigurator<AssignTagsConsumer> consumerConfigurator, IRegistrationContext context)
-    {
-        consumerConfigurator.Options<BatchOptions>(options => options
-            .SetMessageLimit(10)
-            .SetTimeLimit(TimeSpan.FromSeconds(1)));
-    } }
